+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Changed/Fixed
+
+- Update `-v`/`--var` and `-n`/`--namespace` CLI filters to validate that they're given fully qualified or simple symbols, respectively.
+
 ## 2.1.0
 
 Released `2026-08-18`.